requirements.txt
backports.entry-points-selectable==1.1.0
cycler==0.10.0
distlib==0.3.2
filelock==3.0.12
importlib-metadata==4.8.1
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
opencv-python==4.5.3.56
Pillow==8.3.2
platformdirs==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
six==1.16.0
typing-extensions==3.10.0.2
virtualenv==20.7.2
zipp==3.5.0
